Fixes #15915: Move date formatter service to a more common project
