+++ webapp/sources/rudder/rudder-web/src/main/scala/com/normation/rudder/web/comet/AsyncDeployment.scala
@@ -44,11 +44,11 @@ import js._
 import JE._
 import JsCmds._
 import com.normation.rudder.batch._
-import com.normation.rudder.web.components.DateFormaterService
 import org.joda.time.DateTime
 import bootstrap.liftweb.RudderConfig
 import bootstrap.liftweb.RudderConfig.clearCacheService
 import com.normation.rudder.web.model.CurrentUser
+import com.normation.utils.DateFormaterService
 
 class AsyncDeployment extends CometActor with CometListener with Loggable {
